backend/requirements.txt
# Core dependencies
mcp[cli]>=1.0.0
httpx[asyncio]==0.27.0
loguru==0.7.2
python-dotenv==1.0.1
aiohttp>=3.9.0
PyYAML>=6.0.1
fastapi>=0.100.0
websockets>=11.0.3

# Football data providers
understatapi>=0.6.1

# Optional: Database support (future use)
# sqlalchemy>=2.0.0

# Optional: LLM integration (future use)
anthropic>=0.38.0
# openai>=1.54.0

# Optional: Data analysis (future use)
# pandas==2.2.3
# numpy>=1.26.0

# Optional: Task scheduling (future use)
# apscheduler==3.10.4

# Development dependencies
pytest>=8.2.0
pytest-asyncio==0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
litellm>=1.0.0
apscheduler>=3.10.0
